# [AIEX] Make all tables in `AIEXGenFormats.inc` constexpr