--- a/llvm/test/TableGen/CodeGenFormat/FormatTests/MCOperandIndex.td
+++ b/llvm/test/TableGen/CodeGenFormat/FormatTests/MCOperandIndex.td
@@ -105,7 +105,6 @@ def MOV_U20 : TEST_mov20<(outs mMv0Cg20:$reg), (ins uimm20:$imm), [], "mov.u20",
 // CHECK-LABEL: FieldRanges[] = {
 // CHECK: I128_LDA_LDB_NOP_ALU_LNG_NOP_VEC_ALL [[#RANGE0:]]
 // CHECK: // MOV_U20 [[#RANGE1:]]
-// CHECK: // Sentinel [[#RANGE2:]]
 
 // We check below two things:
 // - the (good) mapping between the slots Kinds and the MCOperand Index
@@ -116,13 +115,17 @@ def MOV_U20 : TEST_mov20<(outs mMv0Cg20:$reg), (ins uimm20:$imm), [], "mov.u20",
 //   - op: alu, MI.getOperand(2)
 //   - op: lng, MI.getOperand(3)
 
-// CHECK-LABEL: static MCFormatDesc const Formats[] = {
+// CHECK-LABEL: static constexpr SlotFieldPair const SlotsFields[] = {
+// CHECK: I128_LDA_LDB_NOP_ALU_LNG_NOP_VEC_ALL [[#I128_INDEX]]
+// CHECK-NEXT: { TESTSlotKind::TEST_SLOT_LDA, &FieldsHierarchyTables[[[#%,I128_LDAINDEX:]]] },
+// CHECK-NEXT: { TESTSlotKind::TEST_SLOT_LDB, &FieldsHierarchyTables[[[#%,I128_LDBINDEX:]]] },
+// CHECK-NEXT: { TESTSlotKind::TEST_SLOT_ALU, &FieldsHierarchyTables[[[#%,I128_ALUINDEX:]]] },
+// CHECK-NEXT: { TESTSlotKind::TEST_SLOT_LNG, &FieldsHierarchyTables[[[#%,I128_LNGINDEX:]]] },
+// CHECK-NEXT: { TESTSlotKind::TEST_SLOT_VECA, &FieldsHierarchyTables[[[#%,I128_VECAINDEX:]]] },
+
+// CHECK-LABEL: static constexpr MCFormatDesc const Formats[] = {
 // CHECK: TEST::I128_LDA_LDB_NOP_ALU_LNG_NOP_VEC_ALL - Index : [[#I128_INDEX]]
 // CHECK: /* Slots - Fields mapper */
-// CHECK: TESTSlotKind::TEST_SLOT_LDA, &FieldsHierarchyTables[[[#%,I128_LDAINDEX:]]] },
-// CHECK-SAME: { TESTSlotKind::TEST_SLOT_LDB, &FieldsHierarchyTables[[[#%,I128_LDBINDEX:]]] },
-// CHECK-SAME: { TESTSlotKind::TEST_SLOT_ALU, &FieldsHierarchyTables[[[#%,I128_ALUINDEX:]]] },
-// CHECK-SAME: { TESTSlotKind::TEST_SLOT_LNG, &FieldsHierarchyTables[[[#%,I128_LNGINDEX:]]] },
-// CHECK-SAME: { TESTSlotKind::TEST_SLOT_VECA, &FieldsHierarchyTables[[[#%,I128_VECAINDEX:]]] }}
+// CHECK: llvm::ArrayRef<SlotFieldPair>(&SlotsFields[0], 5),
 // CHECK: /* MCOperand - Slots mapper */
 // CHECK: {&FieldRanges[[[#RANGE0]]], [[#RANGE1]]},