The main goal of this PR is to speed up the build time of the compiler. These are the build time improvements for `AIEXMCFormats.cpp`:
- AIE1: 5.9s => 2.5s
- AIE2: 8.3s => 3s
- AIE2P: 12.8s => 3.9s